refactor: end 커맨드로 끝나는 경우와 King을 잡아서 끝나는 경우, 상태 분리
