+++ src/main/java/chess/domain/state/Checkmate.java
@@ -1,0 +1,17 @@
+package chess.domain.state;
+
+import chess.domain.board.Turn;
+
+public class Checkmate extends Finished{
+
+    private final Turn winner;
+
+    public Checkmate(final Turn turn) {
+        this.winner = turn;
+    }
+
+    @Override
+    public String findWinner() {
+        return winner.convertToColorLabel();
+    }
+}
